feat(react): refactored babel support so options are more easily customized
