+++ package.json
@@ -50,13 +50,13 @@
     "@angular/router": "^9.1.0",
     "@angular/service-worker": "^9.1.0",
     "@angular/upgrade": "^9.1.0",
-    "@babel/core": "7.8.4",
-    "@babel/plugin-proposal-class-properties": "7.8.3",
-    "@babel/plugin-proposal-decorators": "7.8.3",
-    "@babel/plugin-transform-regenerator": "7.8.3",
-    "@babel/preset-env": "7.8.4",
-    "@babel/preset-react": "7.8.3",
-    "@babel/preset-typescript": "7.8.3",
+    "@babel/core": "^7.9.6",
+    "@babel/preset-env": "^7.9.6",
+    "@babel/plugin-proposal-class-properties": "^7.8.3",
+    "@babel/plugin-proposal-decorators": "^7.8.3",
+    "@babel/plugin-transform-regenerator": "^7.8.7",
+    "@babel/preset-typescript": "^7.9.0",
+    "@babel/preset-react": "^7.9.4",
     "@bazel/bazel": "^1.2.0",
     "@bazel/ibazel": "^0.10.3",
     "@cypress/webpack-preprocessor": "^4.1.2",
@@ -73,6 +73,7 @@
     "@ngrx/store-devtools": "9.1.0",
     "@ngtools/webpack": "~9.1.0",
     "@reduxjs/toolkit": "1.3.2",
+    "@rollup/plugin-babel": "5.0.2",
     "@rollup/plugin-commonjs": "11.0.2",
     "@rollup/plugin-image": "2.0.4",
     "@rollup/plugin-node-resolve": "7.1.1",
@@ -110,12 +111,12 @@
     "app-root-path": "^2.0.1",
     "autoprefixer": "9.7.4",
     "axios": "^0.19.0",
-    "babel-loader": "8.0.6",
-    "babel-plugin-const-enum": "^0.0.5",
+    "babel-loader": "^8.1.0",
+    "babel-plugin-const-enum": "^1.0.1",
+    "babel-plugin-macros": "^2.8.0",
+    "babel-plugin-transform-async-to-promises": "^0.8.15",
     "babel-plugin-emotion": "^10.0.29",
-    "babel-plugin-macros": "2.6.1",
     "babel-plugin-styled-components": "^1.10.7",
-    "babel-plugin-transform-async-to-promises": "0.8.15",
     "browserslist": "4.8.7",
     "cacache": "12.0.2",
     "caniuse-lite": "^1.0.30001030",
@@ -126,7 +127,7 @@
     "confusing-browser-globals": "^1.0.9",
     "conventional-changelog-cli": "^2.0.23",
     "copy-webpack-plugin": "5.1.1",
-    "core-js": "^2.6.9",
+    "core-js": "^3.6.5",
     "cosmiconfig": "^4.0.0",
     "css-loader": "3.4.2",
     "cypress": "^4.1.0",
@@ -197,7 +198,6 @@
     "regenerator-runtime": "0.13.3",
     "release-it": "^7.4.0",
     "rollup": "1.31.1",
-    "rollup-plugin-babel": "4.3.3",
     "rollup-plugin-copy": "3.3.0",
     "rollup-plugin-filesize": "6.2.1",
     "rollup-plugin-local-resolve": "1.0.7",
