+++ packages/angular/src/schematics/karma-project/files/src/test.ts__tmpl__
@@ -1,6 +1,5 @@
 // This file is required by karma.conf.js and loads recursively all the .spec and framework files
 <% if (isLibrary) { %>
-import 'core-js/es7/reflect';
 import 'zone.js/dist/zone';
 <% } %>
 import 'zone.js/dist/zone-testing';
