+++ packages/next/src/schematics/application/files/.babelrc__tmpl__
@@ -1,4 +1,4 @@
 {
   "presets": ["next/babel"],
-  "plugins": [<% if (style === 'styled-components') { %>["styled-components", { "pure": true, "ssr": true }]<% } %> ]
+  "plugins": [<% if (style === 'styled-components') { %>["styled-components", { "pure": true, "ssr": true }]<% } %>]
 }
