+++ packages/react/package.json
@@ -31,15 +31,13 @@
     "@nrwl/workspace": "*"
   },
   "dependencies": {
-    "@babel/core": "^7.8.4",
-    "@babel/preset-react": "^7.8.3",
+    "@babel/core": "^7.9.6",
+    "@babel/preset-react": "^7.9.4",
     "@nrwl/cypress": "*",
     "@nrwl/jest": "*",
     "@nrwl/web": "*",
     "@angular-devkit/schematics": "~9.1.0",
     "@svgr/webpack": "^5.2.0",
-    "babel-plugin-emotion": "^10.0.29",
-    "babel-plugin-styled-components": "^1.10.7",
     "confusing-browser-globals": "^1.0.9",
     "eslint-plugin-import": "^2.20.1",
     "eslint-plugin-jsx-a11y": "^6.2.3",
