+++ packages/react/src/schematics/application/files/app/.babelrc__tmpl__
@@ -1,0 +1,9 @@
+{
+  "presets": [
+    "@nrwl/react/babel",
+    <% if (style === '@emotion/styled') { %>"@emotion/babel-preset-css-prop"<% } %>
+  ],
+  "plugins": [
+    <% if (style === 'styled-components') { %>["styled-components", { "pure": true, "ssr": true }]<% } %>
+  ]
+}
