+++ packages/react/src/schematics/application/lib/add-cypress.ts
@@ -1,0 +1,13 @@
+import { externalSchematic, noop, Rule } from '@angular-devkit/schematics';
+import { NormalizedSchema } from '../schema';
+
+export function addCypress(options: NormalizedSchema): Rule {
+  return options.e2eTestRunner === 'cypress'
+    ? externalSchematic('@nrwl/cypress', 'cypress-project', {
+        ...options,
+        name: options.name + '-e2e',
+        directory: options.directory,
+        project: options.projectName,
+      })
+    : noop();
+}
