+++ packages/react/src/schematics/application/lib/normalize-options.ts
@@ -1,0 +1,44 @@
+import { Tree } from '@angular-devkit/schematics';
+import { normalize } from '@angular-devkit/core';
+import { toFileName } from '@nrwl/workspace';
+import { appsDir } from '@nrwl/workspace/src/utils/ast-utils';
+import { NormalizedSchema, Schema } from '../schema';
+import { assertValidStyle } from '../../../utils/assertion';
+
+export function normalizeOptions(
+  host: Tree,
+  options: Schema
+): NormalizedSchema {
+  const appDirectory = options.directory
+    ? `${toFileName(options.directory)}/${toFileName(options.name)}`
+    : toFileName(options.name);
+
+  const appProjectName = appDirectory.replace(new RegExp('/', 'g'), '-');
+  const e2eProjectName = `${appProjectName}-e2e`;
+
+  const appProjectRoot = normalize(`${appsDir(host)}/${appDirectory}`);
+
+  const parsedTags = options.tags
+    ? options.tags.split(',').map((s) => s.trim())
+    : [];
+
+  const fileName = options.pascalCaseFiles ? 'App' : 'app';
+
+  const styledModule = /^(css|scss|less|styl|none)$/.test(options.style)
+    ? null
+    : options.style;
+
+  assertValidStyle(options.style);
+
+  return {
+    ...options,
+    name: toFileName(options.name),
+    projectName: appProjectName,
+    appProjectRoot,
+    e2eProjectName,
+    parsedTags,
+    fileName,
+    styledModule,
+    hasStyles: options.style !== 'none',
+  };
+}
