+++ packages/react/src/schematics/library/library.spec.ts
@@ -1,7 +1,6 @@
 import { Tree } from '@angular-devkit/schematics';
 import { createEmptyWorkspace } from '@nrwl/workspace/testing';
-import { readJsonInTree } from '@nrwl/workspace';
-import { NxJson } from '@nrwl/workspace';
+import { NxJson, readJsonInTree } from '@nrwl/workspace';
 import { runSchematic } from '../../utils/testing';
 
 describe('lib', () => {
@@ -385,7 +384,7 @@ describe('lib', () => {
 
       expect(workspaceJson.projects['my-lib'].architect.build).toMatchObject({
         options: {
-          external: ['react', 'react-dom', 'styled-components'],
+          external: ['react', 'react-dom', 'react-is', 'styled-components'],
         },
       });
     });
@@ -430,7 +429,7 @@ describe('lib', () => {
       });
     });
 
-    it('should add package.json', async () => {
+    it('should add package.json and .babelrc', async () => {
       const tree = await runSchematic(
         'lib',
         {
@@ -441,8 +440,8 @@ describe('lib', () => {
       );
 
       const packageJson = readJsonInTree(tree, '/libs/my-lib/package.json');
-
       expect(packageJson.name).toEqual('@proj/my-lib');
+      expect(tree.exists('/libs/my-lib/.babelrc'));
     });
   });
