+++ packages/react/src/schematics/library/library.ts
@@ -47,6 +47,7 @@ import {
 } from '../../utils/versions';
 import { Schema } from './schema';
 import { libsDir } from '@nrwl/workspace/src/utils/ast-utils';
+import { initRootBabelConfig } from '@nrwl/web/src/utils/rules';
 
 export interface NormalizedSchema extends Schema {
   name: string;
@@ -99,6 +100,7 @@ export default function (schema: Schema): Rule {
         : noop(),
       options.publishable ? updateLibPackageNpmScope(options) : noop(),
       updateAppRoutes(options, context),
+      initRootBabelConfig(),
       formatFiles(options),
     ])(host, context);
   };
