+++ packages/react/src/utils/styled.ts
@@ -1,7 +1,11 @@
 import {
+  babelPluginStyledComponentsVersion,
+  emotionBabelPresetCssPropVersion,
   emotionCoreVersion,
   emotionStyledVersion,
+  reactIsVersion,
   styledComponentsVersion,
+  typesReactIsVersion,
   typesStyledComponentsVersion,
 } from './versions';
 import { PackageDependencies } from './dependencies';
@@ -11,10 +15,13 @@ export const CSS_IN_JS_DEPENDENCIES: {
 } = {
   'styled-components': {
     dependencies: {
+      'react-is': reactIsVersion,
       'styled-components': styledComponentsVersion,
     },
     devDependencies: {
       '@types/styled-components': typesStyledComponentsVersion,
+      '@types/react-is': typesReactIsVersion,
+      'babel-plugin-styled-components': babelPluginStyledComponentsVersion,
     },
   },
   '@emotion/styled': {
@@ -22,6 +29,8 @@ export const CSS_IN_JS_DEPENDENCIES: {
       '@emotion/styled': emotionStyledVersion,
       '@emotion/core': emotionCoreVersion,
     },
-    devDependencies: {},
+    devDependencies: {
+      '@emotion/babel-preset-css-prop': emotionBabelPresetCssPropVersion,
+    },
   },
 };
