+++ packages/react/src/utils/testing.ts
@@ -3,6 +3,7 @@ import { SchematicTestRunner } from '@angular-devkit/schematics/testing';
 import { Rule, Tree } from '@angular-devkit/schematics';
 import { names } from '@nrwl/workspace/src/utils/name-utils';
 import { updateWorkspace } from '@nrwl/workspace/src/utils/workspace';
+import { readJsonInTree } from '@nrwl/workspace';
 
 const testRunner = new SchematicTestRunner(
   '@nrwl/react',
@@ -32,8 +33,47 @@ export function callRule(rule: Rule, tree: Tree) {
   return testRunner.callRule(rule, tree).toPromise();
 }
 
+export function updateNxJson(tree, update: (json: any) => any) {
+  const updated = update(readJsonInTree(tree, '/nx.json'));
+  tree.overwrite('/nx.json', JSON.stringify(updated));
+}
+
 export function createApp(tree: Tree, appName: string): Promise<Tree> {
   const { fileName } = names(appName);
+
+  tree.create(
+    `/apps/${fileName}/src/main.tsx`,
+    `import ReactDOM from 'react-dom';\n`
+  );
+
+  updateNxJson(tree, (json) => {
+    json.projects[appName] = { tags: [] };
+    return json;
+  });
+
+  return callRule(
+    updateWorkspace((workspace) => {
+      workspace.projects.add({
+        name: fileName,
+        root: `apps/${fileName}`,
+        projectType: 'application',
+        sourceRoot: `apps/${fileName}/src`,
+        targets: {},
+      });
+    }),
+    tree
+  );
+}
+
+export function createWebApp(tree: Tree, appName: string): Promise<Tree> {
+  const { fileName } = names(appName);
+
+  tree.create(`/apps/${fileName}/src/index.ts`, `\n`);
+
+  updateNxJson(tree, (json) => {
+    json.projects[appName] = { tags: [] };
+    return json;
+  });
 
   return callRule(
     updateWorkspace((workspace) => {
@@ -52,7 +92,12 @@ export function createApp(tree: Tree, appName: string): Promise<Tree> {
 export function createLib(tree: Tree, libName: string): Promise<Tree> {
   const { fileName } = names(libName);
 
-  tree.create(`/libs/${fileName}/src/index.ts`, `\n`);
+  tree.create(`/libs/${fileName}/src/index.ts`, `import React from 'react';\n`);
+
+  updateNxJson(tree, (json) => {
+    json.projects[libName] = { tags: [] };
+    return json;
+  });
 
   return callRule(
     updateWorkspace((workspace) => {
