+++ packages/storybook/package.json
@@ -32,7 +32,7 @@
   },
   "dependencies": {
     "@nrwl/cypress": "*",
-    "core-js": "^3.2.1",
+    "core-js": "^3.6.5",
     "tree-kill": "1.2.2",
     "ts-loader": "5.3.1",
     "tsconfig-paths-webpack-plugin": "3.2.0",
