+++ packages/web/package.json
@@ -39,28 +39,29 @@
     "@angular-devkit/build-webpack": "~0.901.0",
     "@angular-devkit/core": "~9.1.0",
     "@angular-devkit/schematics": "~9.1.0",
-    "@babel/core": "^7.8.4",
-    "@babel/preset-env": "^7.8.4",
-    "@babel/preset-typescript": "^7.8.3",
+    "@babel/core": "^7.9.6",
+    "@babel/preset-env": "^7.9.6",
     "@babel/plugin-proposal-class-properties": "^7.8.3",
     "@babel/plugin-proposal-decorators": "^7.8.3",
-    "@babel/plugin-transform-regenerator": "^7.8.3",
+    "@babel/plugin-transform-regenerator": "^7.8.7",
+    "@babel/preset-typescript": "^7.9.0",
     "@rollup/plugin-commonjs": "11.0.2",
+    "@rollup/plugin-babel": "5.0.2",
     "@rollup/plugin-image": "2.0.4",
     "@rollup/plugin-node-resolve": "7.1.1",
     "ajv": "6.10.2",
     "autoprefixer": "9.7.4",
-    "babel-loader": "8.0.6",
-    "babel-plugin-const-enum": "0.0.5",
-    "babel-plugin-macros": "2.6.1",
-    "babel-plugin-transform-async-to-promises": "0.8.15",
+    "babel-loader": "^8.1.0",
+    "babel-plugin-const-enum": "^1.0.1",
+    "babel-plugin-macros": "^2.8.0",
+    "babel-plugin-transform-async-to-promises": "^0.8.15",
     "browserslist": "4.8.7",
     "cacache": "12.0.2",
     "caniuse-lite": "^1.0.30001030",
     "circular-dependency-plugin": "5.2.0",
     "clean-css": "4.2.1",
     "copy-webpack-plugin": "5.1.1",
-    "core-js": "^3.2.1",
+    "core-js": "^3.6.5",
     "css-loader": "3.4.2",
     "file-loader": "4.2.0",
     "find-cache-dir": "3.0.0",
@@ -84,13 +85,12 @@
     "rxjs": "^6.5.4",
     "regenerator-runtime": "0.13.3",
     "rollup": "1.31.1",
-    "rollup-plugin-babel": "4.3.3",
-    "rollup-plugin-copy": "3.3.0",
-    "rollup-plugin-filesize": "6.2.1",
-    "rollup-plugin-local-resolve": "1.0.7",
-    "rollup-plugin-peer-deps-external": "2.2.2",
-    "rollup-plugin-postcss": "2.1.1",
-    "rollup-plugin-typescript2": "0.26.0",
+    "rollup-plugin-copy": "^3.3.0",
+    "rollup-plugin-filesize": "^9.0.0",
+    "rollup-plugin-local-resolve": "^1.0.7",
+    "rollup-plugin-peer-deps-external": "^2.2.2",
+    "rollup-plugin-postcss": "^3.1.1",
+    "rollup-plugin-typescript2": "^0.27.1",
     "sass": "1.22.9",
     "sass-loader": "8.0.2",
     "semver": "6.3.0",
