+++ packages/web/src/builders/package/schema.json
@@ -37,7 +37,7 @@
     },
     "babelConfig": {
       "type": "string",
-      "description": "Path to a function which takes a babel config and returns an updated babel config"
+      "description": "(deprecated) Path to a function which takes a babel config and returns an updated babel config"
     },
     "umdName": {
       "type": "string",
