+++ packages/web/src/schematics/application/files/app/.babelrc__tmpl__
@@ -1,0 +1,1 @@
+{}
