+++ packages/web/src/utils/config.spec.ts
@@ -5,6 +5,7 @@ import { LicenseWebpackPlugin } from 'license-webpack-plugin';
 import TsConfigPathsPlugin from 'tsconfig-paths-webpack-plugin';
 import { ProgressPlugin } from 'webpack';
 import { BuildBuilderOptions } from './types';
+import { normalize } from '@angular-devkit/core';
 import CircularDependencyPlugin = require('circular-dependency-plugin');
 import ForkTsCheckerWebpackPlugin = require('fork-ts-checker-webpack-plugin');
 
@@ -19,6 +20,7 @@ describe('getBaseWebpackPartial', () => {
       tsConfig: 'tsconfig.json',
       fileReplacements: [],
       root: '/root',
+      sourceRoot: normalize('/root/src'),
       statsJson: false,
     };
     (<any>(
@@ -164,12 +166,8 @@ describe('getBaseWebpackPartial', () => {
 
       expect(
         (result.module.rules.find((rule) => rule.loader === 'babel-loader')
-          .options as any).presets.find(
-          (p) => p[0].indexOf('@babel/preset-env') !== -1
-        )[1]
-      ).toMatchObject({
-        targets: { esmodules: true },
-      });
+          .options as any).envName
+      ).toMatch('modern');
     });
 
     it('should not override preset-env target for es5', () => {
@@ -177,12 +175,8 @@ describe('getBaseWebpackPartial', () => {
 
       expect(
         (result.module.rules.find((rule) => rule.loader === 'babel-loader')
-          .options as any).presets.find(
-          (p) => p[0].indexOf('@babel/preset-env') !== -1
-        )[1]
-      ).toMatchObject({
-        targets: undefined,
-      });
+          .options as any).envName
+      ).toMatch('legacy');
     });
   });
