+++ packages/web/src/utils/config.ts
@@ -1,6 +1,6 @@
 import * as webpack from 'webpack';
 import { Configuration, ProgressPlugin, Stats } from 'webpack';
-import { dirname, resolve } from 'path';
+import { join, resolve } from 'path';
 import { LicenseWebpackPlugin } from 'license-webpack-plugin';
 import * as CopyWebpackPlugin from 'copy-webpack-plugin';
 import * as TerserWebpackPlugin from 'terser-webpack-plugin';
@@ -8,7 +8,6 @@ import TsConfigPathsPlugin from 'tsconfig-paths-webpack-plugin';
 
 import { AssetGlobPattern, BuildBuilderOptions } from './types';
 import { getOutputHashFormat } from './hash-format';
-import { createBabelConfig } from './babel-config';
 import CircularDependencyPlugin = require('circular-dependency-plugin');
 import ForkTsCheckerWebpackPlugin = require('fork-ts-checker-webpack-plugin');
 
@@ -52,7 +51,10 @@ export function getBaseWebpackPartial(
           loader: `babel-loader`,
           exclude: /node_modules/,
           options: {
-            ...createBabelConfig(dirname(options.main), esm, options.verbose),
+            rootMode: 'upward',
+            cwd: join(options.root, options.sourceRoot),
+            envName: esm ? 'modern' : 'legacy',
+            babelrc: true,
             cacheDirectory: true,
             cacheCompression: false,
           },
@@ -153,7 +155,8 @@ export function createTerserPlugin(esm: boolean, sourceMap: boolean) {
     cache: true,
     sourceMap,
     terserOptions: {
-      ecma: esm ? 6 : 5,
+      ecma: esm ? 8 : 5,
+      // Don't remove safari 10 workaround for ES modules
       safari10: true,
       output: {
         ascii_only: true,
