+++ test.ts
@@ -1,5 +1,5 @@
 // This file is required by karma.conf.js and loads recursively all the .spec and framework files
-import 'core-js/es7/reflect';
+import 'core-js/proposals/reflect-metadata';
 import 'zone.js/dist/zone';
 import 'zone.js/dist/zone-testing';
